# Remove dispatcher classes