@maximelucas 2022-11-02
Nice. I'm not following all the implementation details, but I did think a few days ago that the stats package had *many* classes.